AbstractDirectoryServiceを実装クラスAbstractDirectoryServiceImplに変更し、関連するテストクラスを更新。DiffDirectoryServiceをインターフェースに変更。
