+++ src/test/java/kmg/tool/directorytool/service/MoveDirectoryServiceTest.java
@@ -17,6 +17,7 @@ import org.slf4j.LoggerFactory;
 import ch.qos.logback.classic.Logger;
 import ch.qos.logback.classic.spi.ILoggingEvent;
 import ch.qos.logback.core.read.ListAppender;
+import kmg.tool.directorytool.service.impl.AbstractDirectoryServiceImpl;
 import kmg.tool.directorytool.service.impl.MoveDirectoryServiceImpl;
 
 /**
@@ -38,9 +39,9 @@ public class MoveDirectoryServiceTest extends AbstractDirectoryServiceTest {
      * @return 移動サービスのインスタンス
      */
     @Override
-    protected AbstractDirectoryService createService() {
+    protected AbstractDirectoryServiceImpl createService() {
 
-        final AbstractDirectoryService result = new MoveDirectoryServiceImpl();
+        final AbstractDirectoryServiceImpl result = new MoveDirectoryServiceImpl();
         return result;
 
     }
